Having a change of heart
I think every file that we use, no matter if it's header, or footer, or
file in order, should be put into the work dir. This will probably
lessen the confusion (every file merged in output has to be in
workdir - simple explanation).

Overall the new order in which the files are copied from the workdir is:
1. Header file
2. Files set in order property
3. All other files caught by the regex
4. Footer file